fix ScaleDraw::drawTick() to override Qwt 6.3 API
QwtAbstractScaleDraw::drawTick() changed its tick-length parameter from
int to double in the new Qwt. ScaleDraw still declared int len, making
it a separate overload rather than an override — the base class dispatch
called QwtScaleDraw::drawTick() instead, silently bypassing all custom
logic in ScaleDraw (axis-break filtering, major/minor tick style,
ShowTicksPolicy).

Change the len parameter to double in both the declaration and the
implementation, and add override so future signature drift is caught at
compile time.

+++ qtisas/src/plot2D/ScaleDraw.cpp
@@ -483,7 +483,7 @@ int ScaleDraw::axis() const
 	return a;
 }
 
-void ScaleDraw::drawTick(QPainter *p, double value, int len) const
+void ScaleDraw::drawTick(QPainter *p, double value, double len) const
 {
 	int axis = this->axis();
 	ScaleEngine *sc_engine = (ScaleEngine *)d_plot->axisScaleEngine(axis);
